Внес изменения в проект по рекомендации Любшина Алексея
